requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
qdrant-client==1.12.1
sentence-transformers==3.3.1
torch==2.6.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1